Separate jupyter notebook for classifier from scratch created
To describe a series of successive experiments to improve
the model from scratch, it was decided to create a separate
jupyter notebook.

The experiments will run in the terminal mode. So, code of
model implementation, as well as train an test process, will
be moved into separate .py scripts. The notebook itself only
describes experiment details and contains results.